Skeleton Cocoa Touch Framework
What you changed:

I created the default Cocoa Touch Framework skeleton using XCode Version 8.2.1 (8C1002) and
then modified the Constraid scheme to be shared so that Carthage would be able to build it.

I also stripped out the stupid default comment headers in the files that had the incorrect
copyright information and added the LICENSE.txt file so that people would clearly know that
it is licensed under the MIT license.

I also added a default .gitignore from a ThoughtBot blog post,
https://robots.thoughtbot.com/creating-your-first-ios-framework

Why you made the change:

I did this so that I would have something sane to build this Framework on and get it to a
distributable state.