Remove @testable tag from Constraid import
Why you made the change:

I did this so that only the public interfaces will be exposed to the tests. If I would have
left the @testable tag in place it would have given the testing environment access to non
public interfaces. By removing it, it makes the test suite far more valuable as the test
will fail if we have the incorrect accessibility on a particular interface.