Add support for MacOS
Why you made the change:

I did this so that we would be able to use the framework that we have come to love in iOS
while development MacOS applications as well.

This involved conditionally importing UIKit/AppKit depending on which platform is being
built. It also involved having to typealias the NSLayoutPriority/UILayoutPriority as they
are different dependending on if you are using UIKit/AppKit. It also involved having to
conditionally typealias UIView and NSView so that we could extend them respectively without
having to duplicate a bunch of code which would have sucked in terms of maintenance.