Document the flush(withLeadingMarginOf:) method
Why you made the change:

I did this just so I could get the basic structure of documenting a method after
having read a book the previous night on Swift in code documentation. Clearly
need to document more than the methods. But, I wanted to just get something
started so that I could replicate the pattern across the rest of the files.