Fix flush(withEdge...) where constant didn't behave as expected
Why you made the change:

Specifically, flush(withTrailingEdgeOf:) and flush(withBottomEdgeOf:) were
not resulting in a negative constant at the NSLayoutConstraint level when they
should have if the constant is to conceptually represent an inset. Therefore, I
modified those two methods and the methods that call those methods to adjust
them appropriately so that they all behave as expected now.