Rename centerVertically & centerHorizontally to center
Why you made the change:

I did this so that discoverability would be improved. I know this is a breaking
change and we are already making a number of other breaking changes for massive
discoverability, and understandability improvements for v2.0.0 of this
framework. So, I figured might as well rip the bandaid off and get to a better
place sooner so it can get to more stable faster.