Make center methods return constraint collections
Why you made the change:

I did this so that the collections of the constraints can easily be
stored and can be used to activate/deactivate a set of constraints, or
so that they can be used in animations. I also backfilled test coverage
for these methods. Beyond, that I fixed a bug in one of the methods were
it tried to set the centerXWithinMargins based on centerY.