Add the deprecated flush(withMargins.. methods back
Why you made the change:

I did this so that we wouldn't be taking such a harsh stance for people
using our library. Instead we will give them a major release to deal
with the deprecations and then eliminate them in the major release
following that one. *Note:* I did include the backward compatible
changes to these methods and the bug fixes to these methods as our users
should still receive those in my opinion.