Switch back to GitHub repo referenced images in README.md
Why you made the change:

I did this because I think it should work now that I figured out that
the term slashes in the img tags where what was causing GitHub to not
render the images properly. I know that these increase the size of the
repo a bit but it keeps everything self contained.