Backfill tests for limit(byEdgesOf:) family
Why you made the change:

I did a number of things in this PR.

- Backfilled tests for the limit(byEdgesOf:) family of methods

I did this so that we would have good test coverage.

- Fixed bug with limit(byBottomEdgeOf:) treating constant as a positive

I did this so that the method would behave as expected rather than
non-intuitively going the opposite direction.

- Fixed bug with limit(byTrailingEdgeOf:) treating constant as a positive

I did this so that the method would behave as expected rather than
non-intuitively going the opposite direction.

- Changed limit(byEdgesOf:) family to return collection of constraints

I did this so that people would be able to use the collections to
autolayout based animations easily, etc.