Backfill tests for limit by margin
Why you made the change:

I also changed all the limit by margin methods to return the collection
of constraints. Beyond that, I fixed the two constant bugs in the limit
by trailing margin of and the limit by bottom margin of methods so that
they would behave as expected.