Backfill test for expand(from...) family of functions
Why you made the change:

I did this so that we would have confidence that they are working as
expected and give us confidence to refactor. I also fixed the constant
bugs in the expand(fromBottomEdgeOf:) and expand(fromTrailingEdgeOf:)
methods so that it behaves as expected now. Beyond that I modified the
methods to return the collection of constraints.