Backfilled tests for expand(fromMarginsOf:) family of methods
Why you made the change:

I did this so that we would be able to have confidence in refactoring
this family of methods. I also fixed a couple bugs with the constant of
the fromBottomMarginOf methods and fromTrailingMarginOf method constant
not behaving as expected. I also changed the methods to return the
collection of constraints so that they can easily be managed.