Deprecated expand(fromMarginsOf:, constant:) methods
Why you made the change:

I did this because I have now provided the new more discoverable API
naming with the expand(fromMarginsOf:, offsetBy:) methods. The intent of
this change is to make the API more understandable during the act of
coding without the need to reference documentation or perform a bunch of
trial and error iterations to understand what giving `constant` a
positive value vs a negative value would do.