Rename ConstraidView to View
I did this because by the nature of it being part of a Framework it is
part of a Swift Module making it automatically namespaced under
Constraid. Because of the way imports work with frameworks in Swift you
can refer to Constraid.View if you want to be explicit, or you can
refer to View unless it is ambiguous at which point you have to provide
the Constraid prefix to be explicit.