Move to new non-extension style for flushWithEdges
I did this because the extension approach has already forced certain api
design decisions due to changes in UIView's interface changes over time.
This is bad because it was arbitrary coupling, not coupling because of
an actual dependency on the UIView attributes. So, this approach removes
that unnecessary dependency and future proofs the API design by only
coupling to the attributes of things that are actually needed.