Add tests to show using UILayoutGuide
I did this one to test to make sure that everything works with the
UILayoutGuides. But, also to provide an example. UILayoutGuides should
work as itemB in any of the methods. It will not work however for itemA
as it is of type Constraid.View so that we can set
translatesAutoresizingMaskIntoConstraints to false for that view.