Remove useless constraint casts in tests
I did this so that our tests are a bit cleaner since these actually no
longer do anything. Previously we used to return a collection of
ConstraidConstraint objects. However, now we simply return a collection
of NSLayoutConstraint objects as they are cross-platform. Hence, there
is no longer a reason to cast them to NSLayoutConstraint objects in
tests anymore.