Remove auto activate from manage size methods
I did this because we decided that we didn't want to support auto
activate any longer as it interferes with the AutoLayout animation use
cases. Also, we wanted to promote more use of constraint collection
composition and activation/deactivation.

I also add an example of having to activate in README so people weren't
just clueless about what the expectations where. I also added an example
of constraint collection composition as we wanted to promote that now.