Fix behavior of limit/expand funcs to match concept
I did this because it was extremely confusing what they were actually
doing, and they didn't match the concepts of expand and limit correctly.
Hence, I have reviewed and rewritten all the necessary expand & limit
family of functions so that they now all map correctly to the concepts
of expand and limit.

NOTE: THIS IS A BREAKING CHANGE to existing users of the expand/limit
family of functions.