Remove @discardableResult from all the functions
I did this because for a while now we have had it such that the function
return collections of constraints that by default are NOT activated. The
reason we do this is because we want to support auto layout animations
and more dynamic interface designs.

It turns out that this results in two behaviors.

1. Calling .activate() immediately at the end of the Constraid function
2. Storing the resulting collection of constraints in a variable or
using it in an expresison.

Therefore, there is no longer really the need for the @discardableResult
on each of the functions anymore. They were originally there because the
methods were automatically activating constraints originally. Given that
is not the case we want the compiler to help users remember they have to
do something with the functions result.