Corrected names of proxy methods
I did this because some of the names were following the 'ing' naming
scheme which is generally used when a new instance of an object is being
return and object receiving the method call isn't altered. However, in
this case we are altering the object receiving the method call and
happen to also be returning it to facilitate chaining. So, I changed the
'ing' method names back to non-ing method names.