Clean up acccess levels in ConstraidProxy
I did this so that we wouldn't be exposing the constraint collection or
the base outside of this module, but they would still be accessible to
the methods defined in the extension in different files. Note: If we
were to define them all in the same file as the ConstraidProxy then we
could give them `private`.