Modify some tests to use proxy pattern interface
I did this as it is effectively the primary (outer most) interface we
expect people to use. It then uses the next layer down which are the
individual method interface. These are also part of the public API but
are a slightly less nice API. Modifying the tests to use the proxy
pattern interface also gives us 100% coverage as the proxy pattern
interface uses the individual method interface giving us full coverage.